refactor: renamed modules logger and error handlers
