+++ src/classes/Logger.ts
@@ -1,51 +1,40 @@
-import { NextFunction, Request, Response } from 'express';
-import { finished } from 'stream';
-import fs, { appendFileSync } from 'fs';
-import chalk from 'chalk';
-import path from 'path';
-import { coloredStatusCode } from '../common/coloredSratusCode';
-import { currentTime } from '../common/currentTime';
-
-export const logger = (
-  req: Request,
-  res: Response,
-  next: NextFunction
-): void => {
-  const { method, url, body, query } = req;
-
-  const start = Date.now();
-
-  next();
-
-  finished(res, () => {
-    const end = Date.now();
-    const curTime = currentTime();
-    const { statusCode } = res;
-
-    const strForConsole = `[log] ${curTime} ${chalk.blue(
-      method
-    )} url:${url} query:${JSON.stringify(query)} body:${JSON.stringify(
-      body
-    )} code:${coloredStatusCode(statusCode)} [${end - start}ms]`;
-
-    const strForFile = `${curTime} ${method} url:${url} query:${JSON.stringify(
-      query
-    )} body:${JSON.stringify(body)} code:${statusCode} [${end - start}ms]`;
-
-    process.stdout.write(`${strForConsole}\n`);
-
-    fs.mkdir(path.join(__dirname, '../logs/'), (error) => {
-      if (error && error.code !== 'EEXIST') {
-        process.stderr.write('Catalog have not been created');
-      }
-    });
-    appendFileSync(
-      path.join(__dirname, '../logs/queries.log'),
-      `${strForFile}\n`,
-      {
-        flag: 'a',
-        encoding: 'utf-8',
-      }
-    );
-  });
-};
+// import { createWriteStream } from 'fs';
+// // import chalk from 'chalk';
+// // import path from 'path';
+// // import { coloredStatusCode } from '../helpers/coloredSratusCode';
+// // import { currentTime } from '../helpers/currentTime';
+// // import { serialize } from 'v8';
+// // import { col } from 'sequelize';
+//
+// type SeryalizeObjectType = { [key: string]: string };
+// export const logger = (
+//   file: string
+// ): ((
+//   serializer: (obj: SeryalizeObjectType) => string
+// ) => (kind: string) => (msg: string) => void) => {
+//   const { isTTY } = file;
+//   const stream = isTTY ? file : createWriteStream(file);
+//   return (serializer = logger.defaultSerializer) =>
+//     (kind: string) => {
+//       const color: string = isTTY
+//         ? logger.colors[kind] || logger.colors.info
+//         : '';
+//       const normal = isTTY ? logger.colors.normal : '';
+//
+//       return (msg: string) => {
+//         const record = { msg };
+//         const line = serializer(record);
+//         stream.write(`${color + line + normal}\n`);
+//       };
+//     };
+// };
+//
+// logger.defaultSerializer = (obj: SeryalizeObjectType) =>
+//   Object.values(obj).join('\t');
+//
+// logger.colors = {
+//   warning: '\x1b[1;33m',
+//   error: '\x1b[0;31m',
+//   info: '\x1b[1;37m',
+//   normal: '\x1b[0m',
+// };
